servee/requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.

BeautifulSoup==3.0.8.1
django-improved-inlines==0.1.7a
south==0.7.1
django-treebeard==1.60
django-tagging==0.3.1